chore(bench): put benchmarks in separate crate
