+++ bench/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "bench"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+[dev-dependencies]
+thingbuf = { path = ".." }
+criterion = { version = "0.3.5", features = ["async_tokio"] }
+
+# for comparison benchmarks
+tokio = { version = "1.14.0", features = ["rt", "rt-multi-thread", "sync"] }
+crossbeam = "0.8.1"
+async-std = "1"
+futures = "0.3"
+
+[[bench]]
+name = "sync_mpsc"
+harness = false
+
+[[bench]]
+name = "async_mpsc"
+harness = false
